continuing the removal of gorilla context, updating example to show how current users can deal
