+++ server/server.go
@@ -2,7 +2,6 @@ package server
 
 import (
 	"errors"
-	"fmt"
 	"net/http"
 	"net/http/pprof"
 	"os"
@@ -14,7 +13,6 @@ import (
 
 	"github.com/Sirupsen/logrus"
 	"github.com/go-kit/kit/metrics/provider"
-	"github.com/gorilla/context"
 	"github.com/nu7hatch/gouuid"
 
 	"github.com/NYTimes/gizmo/config/metrics"
@@ -145,29 +143,6 @@ func Stop() error {
 // LogWithFields will feed any request context into a logrus Entry.
 func LogWithFields(r *http.Request) *logrus.Entry {
 	return Log.WithFields(ContextFields(r))
-}
-
-// ContextFields will take a request and convert a context map to logrus Fields.
-func ContextFields(r *http.Request) map[string]interface{} {
-	fields := map[string]interface{}{}
-	for k, v := range context.GetAll(r) {
-		strK := fmt.Sprintf("%+v", k)
-		typeK := fmt.Sprintf("%T-%+v", k, k)
-		// gorilla.mux adds the route to context.
-		// we want to remove it for now
-		if typeK == "mux.contextKey-1" || typeK == "mux.contextKey-0" {
-			continue
-		}
-		// web.varsKey for _all_ mux variables (gorilla or httprouter)
-		if typeK == "web.contextKey-2" {
-			strK = "muxvars"
-		}
-		fields[strK] = fmt.Sprintf("%#+v", v)
-	}
-	fields["path"] = r.URL.Path
-	fields["rawquery"] = r.URL.RawQuery
-
-	return fields
 }
 
 // NewServer will inspect the config and generate
